Refactor code
* Move different EPCs to separate module (sgtin.py, sscc.py)
* Format code with black
